backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
transformers==4.37.2
torch>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10